% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorOutcomeCovariateBuilder.R
\name{createPriorOutcomesCovariateSettings}
\alias{createPriorOutcomesCovariateSettings}
\title{Create settings for adding prior outcomes as covariates}
\usage{
createPriorOutcomesCovariateSettings(outcomeDatabaseSchema = "unknown",
  outcomeTable = "unknown", outcomeIds, outcomeNames, windowStart = -365,
  windowEnd = -1)
}
\arguments{
\item{outcomeDatabaseSchema}{The name of the database schema that is the location
where the data used to define the outcome cohorts is
available.}

\item{outcomeTable}{The tablename that contains the outcome cohorts.}

\item{outcomeIds}{A vector of cohort_definition_ids used to define outcomes}

\item{outcomeNames}{A vector of names of the outcomes, to be used to create
covariate names.}

\item{windowStart}{Start day of the window where covariates are captured,
relative to the index date (0 = index date).}

\item{windowEnd}{End day of the window where covariates are captured,
relative to the index date (0 = index date).}
}
\value{
A covariateSettings object.
}
\description{
Create settings for adding prior outcomes as covariates
}
