% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTablesAndFiguresForReport.R
\name{createTableAndFiguresForReport}
\alias{createTableAndFiguresForReport}
\title{Create figures and tables for report}
\usage{
createTableAndFiguresForReport(outputFolders, databaseNames, maOutputFolder,
  reportFolder)
}
\arguments{
\item{outputFolders}{Vector of names of local folders where the results were generated; make sure 
to use forward slashes (/). D}

\item{databaseNames}{A vector of unique names for the databases.}

\item{maOutputFolder}{A local folder where the meta-anlysis results were be written.}

\item{reportFolder}{A local folder where the tables and figures will be written.}
}
\description{
Create figures and tables for report
}
\details{
This function generates tables and figures for the report on the study results.
}
