% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareResultsForAnalysis.R
\name{prepareResultsForAnalysis}
\alias{prepareResultsForAnalysis}
\title{Prepare results for analysis}
\usage{
prepareResultsForAnalysis(outputFolder, databaseName, maxCores)
}
\arguments{
\item{outputFolder}{Name of local folder where the results were generated; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{databaseName}{A unique name for the database.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Prepare results for analysis
}
\details{
This function generates analyses results, and prepares data for the Shiny app. Requires the study to be executed first.
}
